move `secret access` to `vault.rs`
