+++ relay/src/vault.rs
@@ -4,9 +4,11 @@ use std::path::PathBuf;
 use std::sync::{Arc, RwLock};
 use lazy_static::lazy_static;
 use rusty_vault::core::{Core, SealConfig};
+use rusty_vault::errors::RvError;
+use rusty_vault::logical::{Operation, Request, Response};
 use rusty_vault::storage::{barrier_aes_gcm, physical};
 use serde::{Deserialize, Serialize};
-use serde_json::Value;
+use serde_json::{Map, Value};
 
 #[derive(Serialize, Deserialize, Debug)]
 struct CoreKey {
@@ -92,4 +94,64 @@ fn init() -> CoreInfo {
     }
 
     CoreInfo { core: c, token: root_token }
+}
+
+pub fn read_api(core: &Core, token: &str, path: &str) -> Result<Option<Response>, RvError> {
+    let mut req = Request::new(path);
+    req.operation = Operation::Read;
+    req.client_token = token.to_string();
+    let resp = core.handle_request(&mut req);
+    resp
+}
+
+pub fn write_api(
+    core: &Core,
+    token: &str,
+    path: &str,
+    data: Option<Map<String, Value>>,
+) -> Result<Option<Response>, RvError> {
+    let mut req = Request::new(path);
+    req.operation = Operation::Write;
+    req.client_token = token.to_string();
+    req.body = data;
+
+    let resp = core.handle_request(&mut req);
+    println!("path: {}, req.body: {:?}", path, req.body);
+    resp
+}
+
+/// Write a secret to the vault (k-v)
+pub fn write_secret(name: &str, data: Option<Map<String, Value>>) -> Result<Option<Response>, RvError> {
+    write_api(&CORE.core.read().unwrap(), &CORE.token, &format!("secret/{}", name), data)
+}
+
+/// Read a secret from the vault (k-v)
+pub fn read_secret(name: &str) -> Result<Option<Response>, RvError> {
+    read_api(&CORE.core.read().unwrap(), &CORE.token, &format!("secret/{}", name))
+}
+
+#[cfg(test)]
+mod tests {
+    use serde_json::json;
+    use super::*;
+
+    #[test]
+    fn test_secret() {
+        // create secret
+        let kv_data = json!({
+            "foo": "bar",
+            "id": "oqpXWgEhXa1WDqMWBnpUW4jvrxGqJKVuJATy4MSPdKNS",
+        })
+        .as_object()
+        .unwrap()
+        .clone();
+        write_secret("keyInfo", Some(kv_data.clone())).unwrap();
+
+        let secret = read_secret("keyInfo").unwrap().unwrap().data;
+        assert_eq!(secret, Some(kv_data));
+        println!("secret: {:?}", secret.unwrap());
+
+        assert!(read_secret("foo").unwrap().is_none());
+        assert!(read_api(&CORE.core.read().unwrap(), &CORE.token, "secret1/foo").is_err());
+    }
 }
